+++ src/library/DIPS.Mobile.UI/Components/ListItems/Extensions/NavigationListItem.cs
@@ -1,3 +1,6 @@
+using DIPS.Mobile.UI.Components.ListItems.Options.InLineContent;
+using DIPS.Mobile.UI.Components.ListItems.Options.Title;
+
 namespace DIPS.Mobile.UI.Components.ListItems.Extensions;
 
 [ContentProperty(nameof(InLineContent))]
@@ -5,32 +8,33 @@ public partial class NavigationListItem : ListItem
 {
     private readonly Grid m_contentGrid = new()
     {
-        ColumnDefinitions = new ColumnDefinitionCollection
-        {
-            new (GridLength.Star),
-            new(GridLength.Auto)
-        },
-        RowDefinitions = new RowDefinitionCollection
-        {
-            new (GridLength.Auto)
-        },
+        ColumnDefinitions = new ColumnDefinitionCollection {new(GridLength.Star), new(GridLength.Auto)},
+        RowDefinitions = new RowDefinitionCollection {new(GridLength.Auto)},
         VerticalOptions = LayoutOptions.Center
     };
-    
+
     public NavigationListItem()
     {
-        m_contentGrid.Add(new Image
+        m_contentGrid.Add(
+            new Image
+            {
+                Source = Icons.GetIcon(IconName.arrow_right_s_line),
+                VerticalOptions = LayoutOptions.Center,
+                HorizontalOptions = LayoutOptions.End
+            }, 1);
+
+        TitleOptions = new TitleOptions()
         {
-            Source = Icons.GetIcon(IconName.arrow_right_s_line), 
-            VerticalOptions = LayoutOptions.Center, HorizontalOptions = LayoutOptions.End
-        }, 1);
+            Width = GridLength.Star, LineBreakMode = LineBreakMode.TailTruncation, MaxLines = 1
+        };
+        InLineContentOptions = new InLineContentOptions() {Width = GridLength.Auto};
     }
 
     protected override void OnHandlerChanged()
     {
         base.OnHandlerChanged();
-        
-        if(InLineContent is null)
+
+        if (InLineContent is null)
         {
             AddInLineContent();
         }
@@ -39,7 +43,7 @@ public partial class NavigationListItem : ListItem
     protected override void AddInLineContent()
     {
         var newInLineContent = InLineContent;
-            
+
         m_contentGrid.Insert(0, newInLineContent);
 
         SetInLineContent(m_contentGrid);