+++ src/library/DIPS.Mobile.UI/Components/ListItems/Extensions/RadioButtonListItem.cs
@@ -1,5 +1,7 @@
 using DIPS.Mobile.UI.API.Vibration;
 using DIPS.Mobile.UI.Components.ListItems.Options.Icon;
+using DIPS.Mobile.UI.Components.ListItems.Options.InLineContent;
+using DIPS.Mobile.UI.Components.ListItems.Options.Title;
 using DIPS.Mobile.UI.Components.Selection;
 using SelectionChangedEventArgs = DIPS.Mobile.UI.Components.Selection.SelectionChangedEventArgs;
 
@@ -13,6 +15,9 @@ public partial class RadioButtonListItem : ListItem, ISelectable
     {
         m_iconOptions = new IconOptions();
         IconOptions = m_iconOptions;
+        //Forces the title to take full width
+        TitleOptions = new TitleOptions() {Width = GridLength.Star}; 
+        InLineContentOptions = new InLineContentOptions() {Width = GridLength.Auto};
         
         Command = new Command(() =>
         {